Refactor code related to intereacting with the api endpoint using axios as per TSE feedback
