+++ src/redux/propertiesSlice.js
@@ -1,26 +1,13 @@
 /* eslint-disable no-unused-vars */
 /* eslint-disable no-param-reassign */
-import axios from 'axios';
-import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-import { allPropertiesURL } from './api/apiEndPoints';
-import { axiosDefaults, axiosHeders } from './api/axiosParams';
+import { createSlice } from '@reduxjs/toolkit';
+import { getPropertiesAsync } from './api/apiActions';
 
 const initialState = {
   loading: false,
   properties: [],
   error: '',
 };
-
-export const getPropertiesAsync = createAsyncThunk(
-  'properties/getPropertiesAsync',
-  async () => {
-    await axiosDefaults();
-    await axiosHeders();
-    const response = await axios.get(allPropertiesURL);
-    const properties = await response.data;
-    return properties;
-  },
-);
 
 const propertiesSlice = createSlice({
   name: 'properties',
